feat(daemon): clean the impls of mcp and skill
