+++ crates/daemon/src/config/mcp.rs
@@ -1,0 +1,28 @@
+//! MCP server configuration.
+
+use compact_str::CompactString;
+use serde::{Deserialize, Serialize};
+use std::collections::BTreeMap;
+
+/// MCP server configuration.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct McpServerConfig {
+    /// Server name. If empty, the name will be the command.
+    #[serde(default)]
+    pub name: CompactString,
+    /// Command to spawn.
+    pub command: String,
+    /// Command arguments.
+    #[serde(default)]
+    pub args: Vec<String>,
+    /// Environment variables.
+    #[serde(default)]
+    pub env: BTreeMap<String, String>,
+    /// Auto-restart on failure.
+    #[serde(default = "default_true")]
+    pub auto_restart: bool,
+}
+
+fn default_true() -> bool {
+    true
+}
